feature: move parsing functions to separate file
